- [x] date issue fixes along with existing identifier - [x] activity test and fixes - [x] test cases for valid data and fixes for activity, result and indicator - [x] testing flow wip - [x] sheet and header checks

+++ app/Http/Controllers/Admin/ImportActivity/ImportXlsController.php
@@ -9,11 +9,6 @@ use App\Http\Requests\Activity\UploadActivity\ImportXlsRequest;
 use App\IATI\Services\Activity\ActivityService;
 use App\IATI\Services\ImportActivity\ImportXlsService;
 use App\IATI\Services\ImportActivityError\ImportActivityErrorService;
-use App\XlsImporter\Foundation\Mapper\Activity;
-use App\XlsImporter\Foundation\Mapper\Indicator;
-use App\XlsImporter\Foundation\Mapper\Period;
-use App\XlsImporter\Foundation\Mapper\Result;
-use Arr;
 use Exception;
 use Illuminate\Contracts\View\View;
 use Illuminate\Database\DatabaseManager;
@@ -78,63 +73,6 @@ class ImportXlsController extends Controller
     public function index(): View|JsonResponse|RedirectResponse
     {
         try {
-            // dd(awsHasFile('XlsImporter/tmp/1/2/status.json'));
-            // $elements = readElementJsonSchema();
-            // $final = [];
-
-            // foreach ($elements as $elementName => $content) {
-            //     $attributes = $this->getAttributes(Arr::get($content, 'attributes', []));
-            //     $subElements = $this->getSubElements(Arr::get($content, 'sub_elements', []));
-
-            //     foreach ($subElements as $key => $value) {
-            //         $attributes[$key] = $value;
-            //     }
-
-            //     $final[$elementName] = $attributes;
-            // }
-
-            // dd(json_encode($final));
-
-            // $this->getLinearizedElement();
-
-            // $data = getCodeList('FileFormat', 'Activity', false);
-            // $data = json_encode(getCodeList('Region', 'Activity'));
-            // test
-            // $data = file_get_contents(app_path() . '/XlsImporter/Templates/test.json');
-            // $activityMapper = new Activity();
-            // $activityMapper->map($data);
-            // $org_id = Auth::user()->organization->id;
-            // $xlsType = "indicator";
-            // dd($this->importXlsService->dbIatiIdentifiers($org_id, $xlsType));
-
-            // period
-
-            // $data = file_get_contents(app_path() . '/XlsImporter/Templates/period.json');
-            // $data = json_decode($data, true, 512, 0);
-            // $periodMapper = new Period();
-            // $periodMapper->map($data);
-
-            // indicator
-
-            // $data = file_get_contents(app_path() . '/XlsImporter/Templates/indicator.json');
-            // $data = json_decode($data, true, 512, 0);
-            // $indicatorMapper = new Indicator();
-            // $indicatorMapper->map($data);
-            $data = file_get_contents(app_path('/XlsImporter/Templates/result.json'));
-            $data = json_decode($data, true, 512, 0);
-            $resultMapper = new Result();
-            $resultMapper->map($data);
-
-            //            $data = file_get_contents(app_path() . '/XlsImporter/Templates/period.json');
-//            $periodMapper = new Period();
-//            $periodMapper->map($data);
-
-            // $periodMapper
-
-            //            $resultData = file_get_contents(app_path('/XlsImporter/Templates/result.json'));
-//            $resultMapper = new Result();
-//            $resultMapper->map($resultData);
-
             if (!Auth::user()->organization_id) {
                 Session::put('error', 'User is not associated with any organization.');
 
@@ -143,7 +81,6 @@ class ImportXlsController extends Controller
 
             return view('admin.import.xls.index');
         } catch (Exception $e) {
-            dd($e);
             logger()->error($e->getMessage());
 
             return response()->json(['success' => false, 'error' => 'Error has occurred while rendering activity import page.']);
@@ -170,7 +107,6 @@ class ImportXlsController extends Controller
 
             return response()->json(['success' => true, 'message' => 'Uploaded successfully']);
         } catch (Exception $e) {
-            logger()->error($e);
             logger()->error($e->getMessage());
 
             return response()->json(['success' => false, 'error' => 'Error has occurred while rendering activity import page.']);
@@ -188,29 +124,27 @@ class ImportXlsController extends Controller
     public function importValidatedActivities(Request $request): mixed
     {
         try {
-            // $this->importXlsService->deleteImportStatus();
             $this->db->beginTransaction();
             $status = $this->importXlsService->getImportStatus();
+            $activities = $request->get('activites');
 
-            // dd($status);
-            $xlsType = $status['template'];
-            $activities = $request->query('activities');
-
-            if ($activities) {
-                $activities = json_decode($request->query('activities'));
-
-                $this->importXlsService->create($activities, $xlsType);
+            if (empty($status)) {
+                return response()->json(['success' => false, 'message' => 'Please ensure that you have uploaded xls file.']);
             }
 
+            if (empty($activities)) {
+                return response()->json(['success' => false, 'message' => 'Please select the data you want to add.']);
+            }
+
+            $xlsType = $status['template'];
+            $this->importXlsService->create($activities, $xlsType);
             $this->importXlsService->deleteImportStatus();
 
             $this->db->commit();
 
             return response()->json(['success' => true, 'message' => 'Imported successfully']);
         } catch (Exception $e) {
-            dd($e);
-            Session::put('error', 'Error occurred while importing activity');
-            logger()->error($e->getMessage());
+            logger()->error($e);
 
             return redirect()->back()->withResponse(['success' => false, 'message' => 'Error has occurred while importing activity.']);
         }
@@ -238,11 +172,8 @@ class ImportXlsController extends Controller
     {
         try {
             $status = $this->importXlsService->getImportStatus();
-            // $filetype = Session::get('import_filetype');
 
             if (empty($status)) {
-                Session::put('error', 'Please upload xls file to import activity.');
-
                 return response()->json(['status' => 'error', 'message' => 'Please upload xls file to import activity.']);
             }
